refactor(backup): :truck: rename backup test support module to utils
