+++ src-tauri/src/backup/tests/archive.rs
@@ -1,4 +1,4 @@
-use super::support::{ConfigFileGuard, lock_config_file};
+use super::utils::{ConfigFileGuard, lock_config_file};
 use crate::backup::archive::{
     ZIP_COMMENT_LOCAL_TIME_MARKER, ZipTimestampInterpretation, add_directory, compress_to_file,
     decompress_from_file, local_result_to_timestamp, system_time_to_zip_datetime,
