+++ src-tauri/src/backup/tests/game.rs
@@ -1,4 +1,4 @@
-use super::support::{ConfigFileGuard, lock_config_file};
+use super::utils::{ConfigFileGuard, lock_config_file};
 use crate::backup::archive::system_time_to_zip_datetime;
 use crate::backup::state_fingerprint::{fingerprint_source_state, fingerprint_zip_state};
 use crate::backup::{
